Delay many re.compile calls.
This is done either by moving the compile calls out of module scope,
or by importing code lazily.

This takes re.compile from 11% of startup time, and 100s of calls,
to 3.5% of startup time, and tens of calls.
