+++ tests/python/pants_test/backend/core/tasks/test_markdown_to_html.py
@@ -13,7 +13,7 @@ import bs4
 import mock
 
 from pants.backend.core.targets.doc import Page
-from pants.backend.core.tasks import markdown_to_html
+from pants.backend.core.tasks import markdown_to_html_utils
 from pants.backend.core.tasks.markdown_to_html import MarkdownToHtml
 from pants.base.exceptions import TaskError
 from pants_test.tasks.task_test_base import TaskTestBase
@@ -27,76 +27,76 @@ charlie"""
 class ChooseLinesTest(unittest.TestCase):
   def test_include_no_params(self):
     self.assertEquals(
-        markdown_to_html.choose_include_text(ABC, '', 'fake.md'),
+        markdown_to_html_utils.choose_include_text(ABC, '', 'fake.md'),
         '\n'.join(['able', 'baker', 'charlie']))
 
   def test_include_start_at(self):
     self.assertEquals(
-        markdown_to_html.choose_include_text(ABC, 'start-at=abl', 'fake.md'),
+        markdown_to_html_utils.choose_include_text(ABC, 'start-at=abl', 'fake.md'),
         '\n'.join(['able', 'baker', 'charlie']))
 
     self.assertEquals(
-        markdown_to_html.choose_include_text(ABC, 'start-at=bak', 'fake.md'),
+        markdown_to_html_utils.choose_include_text(ABC, 'start-at=bak', 'fake.md'),
         '\n'.join(['baker', 'charlie']))
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-at=xxx', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-at=xxx', 'fake.md'),
       '')
 
   def test_include_start_after(self):
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-after=bak', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-after=bak', 'fake.md'),
       'charlie')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-after=cha', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-after=cha', 'fake.md'),
       '')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-after=xxx', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-after=xxx', 'fake.md'),
       '')
 
   def test_include_end_at(self):
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-at=abl', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-at=abl', 'fake.md'),
       'able')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-at=bak', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-at=bak', 'fake.md'),
       '\n'.join(['able', 'baker']))
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-at=xxx', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-at=xxx', 'fake.md'),
       '')
 
   def test_include_end_before(self):
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-before=abl', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-before=abl', 'fake.md'),
       '')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-before=xxx', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-before=xxx', 'fake.md'),
       '')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'end-before=bak', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'end-before=bak', 'fake.md'),
       'able')
 
   def test_include_start_at_end_at(self):
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-at=abl&end-at=abl', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-at=abl&end-at=abl', 'fake.md'),
       'able')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-at=cha&end-at=cha', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-at=cha&end-at=cha', 'fake.md'),
       'charlie')
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-at=abl&end-at=bak', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-at=abl&end-at=bak', 'fake.md'),
       '\n'.join(['able', 'baker']))
 
     self.assertEquals(
-      markdown_to_html.choose_include_text(ABC, 'start-at=bak&end-at=abl', 'fake.md'),
+      markdown_to_html_utils.choose_include_text(ABC, 'start-at=bak&end-at=abl', 'fake.md'),
       '')
